+++ packages/widget-attributions/src/index.js
@@ -1,0 +1,3 @@
+import AttributionsWidget from './AttributionsWidget.vue';
+
+export default AttributionsWidget;