+++ packages/widget-passage-citation-toc/src/index.js
@@ -1,0 +1,9 @@
+import PassageCitationTOCWidget, {
+  displayNameCallback,
+} from './PassageCitationTOCWidget.vue';
+
+export default PassageCitationTOCWidget;
+
+// NOTE: exporting `displayNameCallback` allows us to shim the
+// dynamic display name functionality into SV 1
+export { displayNameCallback };