+++ src/main.js
@@ -16,6 +16,7 @@ import { SkeletonPlugin } from '@scaife-viewer/skeleton';
 
 import { DefaultModeReader } from '@scaife-viewer/widget-reader';
 
+import AlignmentsModeReader from '@scaife-viewer/reader-alignments-mode';
 import ImageModeReader, {
   iconMap as imageModeReaderIconMap,
 } from '@scaife-viewer/reader-image-mode';
@@ -30,8 +31,6 @@ import { iconMap as commonIconMap } from '@scaife-viewer/common';
 import { iconMap as audioIconMap } from '@scaife-viewer/widget-audio';
 import { iconMap as neIconMap } from '@scaife-viewer/widget-named-entities';
 import { iconMap as deIconMap } from '@scaife-viewer/widget-dictionary-entries';
-
-import AlignmentsModeReader from '../packages/reader-alignments-mode/src';
 
 import App from './App.vue';
 import store, { apolloProvider } from './store';
@@ -63,6 +62,9 @@ Vue.use(SkeletonPlugin, {
       [DISPLAY_MODE_INTERLINEAR]: InterlinearModeReader,
       [DISPLAY_MODE_SYNTAX_TREES]: SyntaxTreesModeReader,
       [DISPLAY_MODE_DEFAULT]: DefaultModeReader,
+    },
+    citationSchemeMap: {
+      book: 'Books',
     },
     pageTitle: title => `SV Demo | ${title}`,
   },